requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic>=2.7
httpx>=0.27
python-dotenv>=1.0
litellm>=1.40.0
tqdm>=4.66.1
